requirements_pdf_chatting.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromdb